Image Recognition Project
This is a small project which identifies that the person in the camera is me or one of my friend. We tested this program even with and without specs to see whether it works properly or not. We were delighted to find it working perfectly. I came to know that the more information (images) you put inside, the more accurate results are shows in this project. 

for this project I have used :
Epochs            - 100
Batch Size       - 64 
Learning Rate - 0.003